Allow empty sidebar, correctly check mixed content
