+++ src/components/AppSidebar/AppSidebar.vue
@@ -20,9 +20,6 @@
   -
   -->
 
-<!-- Follows the tab aria guidelines
-	https://www.w3.org/TR/wai-aria-practices/examples/tabs/tabs-1/tabs.html -->
-
 <docs>
 
 	### General description
@@ -82,11 +79,25 @@
 		}
 	</script>
 	```
+	```
+
+	### Empty sidebar for e.g. empty content component.
+	```vue
+	<template>
+		<AppSidebar
+			title="cat-picture.jpg"
+			:empty="true">
+			<EmptyContent icon="icon-search">
+				Content not found.
+			</EmptyContent>
+		</AppSidebar>
+	</template>
+	```
 
 </docs>
 
 <template>
-	<transition name="slide-right">
+	<transition name="slide-right" appear>
 		<aside id="app-sidebar-vue" class="app-sidebar">
 			<header :class="{
 					'app-sidebar-header--with-figure': hasFigure,
@@ -100,7 +111,7 @@
 					@click.prevent="closeSidebar" />
 
 				<!-- sidebar header illustration/figure -->
-				<div v-if="hasFigure"
+				<div v-if="hasFigure && !empty"
 					:class="{
 						'app-sidebar-header__figure--with-action': hasFigureClickListener
 					}"
@@ -113,7 +124,7 @@
 				</div>
 
 				<!-- sidebar details -->
-				<div :class="{ 'app-sidebar-header__desc--with-star': canStar, 'app-sidebar-header__desc--with-subtitle': subtitle && !titleEditable, 'app-sidebar-header__desc--editable': titleEditable && !subtitle, 'app-sidebar-header__desc--with-subtitle--editable': titleEditable && subtitle}" class="app-sidebar-header__desc">
+				<div v-if="!empty" :class="{ 'app-sidebar-header__desc--with-star': canStar, 'app-sidebar-header__desc--with-subtitle': subtitle && !titleEditable, 'app-sidebar-header__desc--editable': titleEditable && !subtitle, 'app-sidebar-header__desc--with-subtitle--editable': titleEditable && subtitle}" class="app-sidebar-header__desc">
 					<!-- favourite icon -->
 					<a v-if="canStar"
 						:class="{ 'icon-starred': isStarred&& !starLoading, 'icon-star': !isStarred && !starLoading, 'icon-loading-small': starLoading }"
@@ -151,60 +162,29 @@
 						<slot name="secondary-actions" />
 					</Actions>
 				</div>
-				<div v-if="$slots['primary-actions']" class="app-sidebar-header__action">
+				<div v-if="$slots['primary-actions'] && !empty" class="app-sidebar-header__action">
 					<slot name="primary-actions" />
 				</div>
 			</header>
 
-			<!-- tabs navigation -->
-			<nav v-if="hasMultipleTabs"
-				class="app-sidebar-tabs__nav"
-				@keydown.left.exact.prevent="focusPreviousTab"
-				@keydown.right.exact.prevent="focusNextTab"
-				@keydown.tab.exact.prevent="focusActiveTabContent"
-				@keydown.page-up.exact.prevent="focusFirstTab"
-				@keydown.page-down.exact.prevent="focusLastTab">
-				<ul>
-					<li v-for="tab in tabs" :key="tab.id" class="app-sidebar-tabs__tab">
-						<a :id="tab.id"
-							:aria-controls="`tab-${tab.id}`"
-							:aria-selected="activeTab === tab.id"
-							:class="{ active: activeTab === tab.id }"
-							:data-id="tab.id"
-							:href="`#tab-${tab.id}`"
-							:tabindex="activeTab === tab.id ? null : -1"
-							role="tab"
-							@click.prevent="setActive">
-							<span :class="tab.icon" class="app-sidebar-tabs__tab-icon" />
-							{{ tab.name }}
-						</a>
-					</li>
-				</ul>
-			</nav>
-
-			<!-- tabs content -->
-			<div :class="{'app-sidebar-tabs__content--multiple': hasMultipleTabs}"
-				class="app-sidebar-tabs__content">
-				<slot :active-tab="activeTab" />
-			</div>
+			<AppSidebarTabs>
+				<slot :active="active" />
+			</AppSidebarTabs>
 		</aside>
 	</transition>
 </template>
 
 <script>
-import Vue from 'vue'
 import Actions from '../Actions'
 import Focus from '../../directives/Focus'
 import l10n from '../../mixins/l10n'
-
-const IsValidString = function(value) {
-	return value && typeof value === 'string' && value.trim() !== '' && value.indexOf(' ') === -1
-}
+import AppSidebarTabs from '../AppSidebarTabs/AppSidebarTabs'
 
 export default {
 	name: 'AppSidebar',
 	components: {
 		Actions,
+		AppSidebarTabs,
 	},
 	directives: {
 		focus: Focus,
@@ -267,6 +247,16 @@ export default {
 		},
 
 		/**
+		 * Only display close button and default slot content.
+		 * Don't display other header content and primary and secondary actions.
+		 * Useful when showing the EmptyContent component as content.
+		 */
+		empty: {
+			type: Boolean,
+			default: false,
+		},
+
+		/**
 		 * Force the actions to display in a three dot menu
 		 */
 		forceMenu: {
@@ -277,10 +267,7 @@ export default {
 
 	data() {
 		return {
-			tabs: [],
-			activeTab: '',
 			isStarred: this.starred,
-			children: [],
 		}
 	},
 
@@ -291,41 +278,15 @@ export default {
 		hasFigure() {
 			return this.$slots['header'] || this.background
 		},
-		hasMultipleTabs() {
-			return this.tabs.length > 1
-		},
 		hasFigureClickListener() {
 			return this.$listeners['figure-click']
-		},
-		currentTabIndex() {
-			return this.tabs.findIndex(tab => tab.id === this.activeTab)
 		},
 	},
 
 	watch: {
-		active(active) {
-			// prevent running it twice
-			if (active !== this.activeTab) {
-				this.updateActive()
-			}
-		},
 		starred() {
 			this.isStarred = this.starred
 		},
-
-		// update the tabs list if the children
-		// length change
-		children() {
-			this.updateTabs()
-		},
-	},
-
-	mounted() {
-		// init the tabs list
-		this.updateTabs()
-
-		// let's make the children list reactive
-		this.children = this.$children
 	},
 
 	methods: {
@@ -346,125 +307,12 @@ export default {
 		},
 
 		/**
-		 * Set the current active tab
-		 */
-		setActive({ target }) {
-			// if click on icon, make sure we get the link
-			const id = target.closest('a').dataset.id
-			this.activeTab = id
-			this.$emit('update:active', id)
-
-		},
-
-		/**
-		 * Focus the previous tab
-		 * and emit to the parent component
-		 */
-		focusPreviousTab() {
-			if (this.currentTabIndex > 0) {
-				this.activeTab = this.tabs[this.currentTabIndex - 1].id
-				this.$emit('update:active', this.activeTab)
-			}
-			this.focusActiveTab() // focus nav item
-		},
-
-		/**
-		 * Focus the next tab
-		 * and emit to the parent component
-		 */
-		focusNextTab() {
-			if (this.currentTabIndex < this.tabs.length - 1) {
-				this.activeTab = this.tabs[this.currentTabIndex + 1].id
-				this.$emit('update:active', this.activeTab)
-			}
-			this.focusActiveTab() // focus nav item
-		},
-
-		/**
-		 * Focus the first tab
-		 * and emit to the parent component
-		 */
-		focusFirstTab() {
-			this.activeTab = this.tabs[0].id
-			this.$emit('update:active', this.activeTab)
-			this.focusActiveTab() // focus nav item
-		},
-
-		/**
-		 * Focus the last tab
-		 * and emit to the parent component
-		 */
-		focusLastTab() {
-			this.activeTab = this.tabs[this.tabs.length - 1].id
-			this.$emit('update:active', this.activeTab)
-			this.focusActiveTab() // focus nav item
-		},
-
-		/**
-		 * Focus the current active tab
-		 */
-		focusActiveTab() {
-			this.$el.querySelector('#' + this.activeTab).focus()
-		},
-
-		/**
-		 * Focus the content on tab
-		 * see aria accessibility guidelines
-		 */
-		focusActiveTabContent() {
-			this.$el.querySelector('#tab-' + this.activeTab).focus()
-		},
-
-		/**
-		 * Update the current active tab
-		 */
-		updateActive() {
-			this.activeTab = this.active
-				&& this.tabs.findIndex(tab => tab.id === this.active) !== -1
-				? this.active
-				: this.tabs.length > 0
-					? this.tabs[0].id
-					: ''
-		},
-
-		/**
 		 * Toggle the favourite state
 		 * and emit to the parent component
 		 */
 		toggleStarred() {
 			this.isStarred = !this.isStarred
 			this.$emit('update:starred', this.isStarred)
-		},
-
-		/**
-		 * Manually update the sidebar tabs according to $children
-		 */
-		updateTabs() {
-			// Init tabs from $children
-			const tabs = this.$children.filter(child =>
-				(child.name && typeof child.name === 'string')
-				&& IsValidString(child.id)
-				&& IsValidString(child.icon)
-			)
-
-			// tabs are optional, but you can use either tabs or non-tab-content only
-			if (tabs.length !== 0 && tabs.length !== this.$children.length) {
-				Vue.util.warn(`Mixing tabs and non-tab-content is not possible.`)
-			}
-
-			this.tabs = tabs.sort((a, b) => {
-				const orderA = a.order || 0
-				const orderB = b.order || 0
-				if (orderA === orderB) {
-					return OC.Util.naturalSortCompare(a.name, b.name)
-				}
-				return orderA - orderB
-			})
-
-			// init active tab if exists
-			if (this.tabs.length > 0) {
-				this.updateActive()
-			}
 		},
 
 		/**
@@ -691,85 +539,6 @@ $top-buttons-spacing: 6px;
 						margin-top: 0;
 					}
 				}
-			}
-		}
-	}
-	.app-sidebar-tabs {
-		&__nav {
-			margin-top: 10px;
-			ul {
-				display: flex;
-				justify-content: stretch;
-			}
-		}
-		&__tab {
-			display: block;
-			flex: 1 1;
-			min-width: 0;
-			text-align: center;
-			a {
-				position: relative;
-				display: block;
-				overflow: hidden;
-				padding: 25px 5px 5px 5px;
-				transition: color var(--animation-quick), opacity var(--animation-quick), border-color var(--animation-quick);
-				text-align: center;
-				white-space: nowrap;
-				text-overflow: ellipsis;
-				opacity: $opacity_normal;
-				color: var(--color-main-text);
-				border-bottom: 1px solid var(--color-border);
-
-				&:hover,
-				&:focus,
-				&:active,
-				&.active {
-					opacity: $opacity_full;
-					.app-sidebar-tabs__tab-icon {
-						opacity: $opacity_full;
-					}
-				}
-				&:not(.active):hover,
-				&:not(.active):focus {
-					border-bottom-color: var(--color-background-darker);
-					box-shadow: inset 0 -1px 0 var(--color-background-darker);
-				}
-				&.active {
-					color: var(--color-text-light);
-					border-bottom-color: var(--color-text-light);
-					box-shadow: inset 0 -1px 0 var(--color-text-light);
-					font-weight: bold;
-				}
-				// differentiate the two for accessibility purpose
-				// make sure the user knows she's focusing the navigation
-				// and can use arrows/home/pageup...
-				&:focus {
-					border-bottom-color: var(--color-primary-element);
-					box-shadow: inset 0 -1px 0 var(--color-primary-element);
-				}
-			}
-		}
-
-		&__tab-icon {
-			position: absolute;
-			top: 0;
-			left: 0;
-			width: 100%;
-			height: 25px;
-			transition: opacity var(--animation-quick);
-			opacity: $opacity_normal;
-			background-position: center 8px;
-			background-size: 16px;
-		}
-
-		&__content {
-			position: relative;
-			// take full available height
-			flex: 1 1 100%;
-			// force the use of the tab component if more than one tab
-			// you can just put raw content if you don't use tabs
-			&--multiple > :not(section) {
-				display: none;
 			}
 		}
 	}
