+++ tests/OC.js
@@ -1,15 +1,20 @@
-export class OC {
+export default {
 
 	getLanguage() {
 		return 'en-GB'
-	}
+	},
 
 	getLocale() {
 		return 'en_GB'
-	}
+	},
 
 	isUserAdmin() {
 		return true
-	}
+	},
 
+	Util: {
+		naturalSortCompare(a, b) {
+			return 0
+		}
+	}
 }
