+++ tests/setup.js
@@ -20,9 +20,9 @@
  *
  */
 
-import { OC } from './OC.js'
+import OC from './OC.js'
 
-global.OC = new OC()
+global.OC = OC
 
 global.TRANSLATIONS = []
 global.SCOPE_VERSION = 1
